Refactor Ugin's Conjurant, Protean Hydra with a new ability
